+++ lib/td_dq_web/views/rule_result_view.ex
@@ -1,10 +1,12 @@
 defmodule TdDqWeb.RuleResultView do
   use TdDqWeb, :view
 
-  alias TdDqWeb.RuleResultView
+  def render("show.json", %{rule_result: rule_result}) do
+    %{data: render_one(rule_result, __MODULE__, "rule_result.json")}
+  end
 
   def render("index.json", %{rule_results: rule_results}) do
-    %{data: render_many(rule_results, RuleResultView, "rule_result.json")}
+    %{data: render_many(rule_results, __MODULE__, "rule_result.json")}
   end
 
   def render("rule_result.json", %{rule: %{} = rule} = assigns) do
@@ -21,7 +23,18 @@ defmodule TdDqWeb.RuleResultView do
 
   def render("rule_result.json", %{rule_result: rule_result}) do
     rule_result
-    |> Map.take([:id, :implementation_key, :date, :result, :records, :errors, :result_type])
+    |> Map.take([
+      :id,
+      :implementation_key,
+      :date,
+      :result,
+      :records,
+      :errors,
+      :result_type,
+      :execution_id
+    ])
+    |> Enum.reject(fn {_, v} -> is_nil(v) end)
+    |> Map.new()
     |> with_params(rule_result)
   end