+++ test/td_dq_web/controllers/search_controller_test.exs
@@ -42,43 +42,5 @@ defmodule TdDqWeb.SearchControllerTest do
                |> post(Routes.search_path(conn, :search_rules))
                |> json_response(:ok)
     end
-
-    @tag authentication: [role: "admin"]
-    test "admin can search implementations", %{conn: conn} do
-      assert %{"data" => [_]} =
-               conn
-               |> post(Routes.search_path(conn, :search_implementations))
-               |> json_response(:ok)
-    end
-
-    @tag authentication: [role: "user"]
-    test "user with permissions can search implementations", %{
-      conn: conn,
-      claims: %{user_id: user_id},
-      domain: %{id: domain_id}
-    } do
-      assert %{"data" => [], "user_permissions" => perms} =
-               conn
-               |> post(Routes.search_path(conn, :search_implementations))
-               |> json_response(:ok)
-
-      assert %{"execute" => false, "manage" => false} = perms
-
-      create_acl_entry(user_id, "domain", domain_id, [
-        :view_quality_rule,
-        :manage_quality_rule_implementations
-      ])
-
-      create_acl_entry(user_id, "business_concept", @business_concept_id, [
-        :execute_quality_rule_implementations
-      ])
-
-      assert %{"data" => [_], "user_permissions" => perms} =
-               conn
-               |> post(Routes.search_path(conn, :search_implementations))
-               |> json_response(:ok)
-
-      assert %{"execute" => true, "manage" => true} = perms
-    end
   end
 end